Move MantleTx to its own module.
