+++ core/src/block/genesis.rs
@@ -10,11 +10,12 @@ use crate::{
     block::{Block, BlockTransactions},
     header::Header,
     mantle::{
-        MantleTx, Note, Op, OpProof, SignedMantleTx,
+        Note, Op, OpProof, SignedMantleTx,
         ledger::{BoundedOutputs, Inputs, Outputs},
         ops::{channel::inscribe::InscriptionOp, sdp::SDPDeclareOp, transfer::TransferOp},
         transactions::{
-            GenesisTx, MAX_OPS_PER_TX, Ops, VerificationError, genesis_tx, tx::OpsProofs,
+            GenesisTx, MAX_OPS_PER_TX, Ops, OpsProofs, VerificationError, genesis_tx,
+            mantle_tx::MantleTx,
         },
     },
 };
