+++ core/src/mantle/mod.rs
@@ -11,8 +11,8 @@ pub mod transactions;
 pub use gas::{GasCalculator, GasConstants};
 pub use ledger::{Note, NoteId, Utxo, Value};
 pub use ops::{Op, OpProof};
-pub use transactions::{CryptarchiaParameter, GenesisTime, hash::TxHash};
+pub use transactions::{CryptarchiaParameter, GenesisTime, hash::TxHash, mantle_tx::MantleTx};
 
-pub use crate::mantle::transactions::{MantleTx, SignedMantleTx, VerificationError};
+pub use crate::mantle::transactions::{SignedMantleTx, VerificationError};
 
 pub const MAX_MANTLE_TXS: usize = 1024;
