+++ core/src/mantle/transactions/codec.rs
@@ -4,11 +4,11 @@ use nom::IResult;
 
 use crate::{
     mantle::{
-        MantleTx, Op, SignedMantleTx,
+        Op, SignedMantleTx,
         nom::{NomDecode as _, NomEncode as _},
         ops::codec::{decode_ops_proofs, encode_ops_proofs},
         transactions::{
-            MantleTxGasContext,
+            mantle_tx::{MantleTx, MantleTxGasContext},
             states::{Unverified, VerificationState},
         },
     },
@@ -117,7 +117,7 @@ mod tests {
                 transfer::TransferOp,
             },
             traits::Hashable as _,
-            transactions::{GasPrices, Ops, tx::OpsProofs},
+            transactions::{GasPrices, Ops, OpsProofs},
         },
         proofs::{
             channel_multi_sig_proof::{ChannelMultiSigProof, IndexedSignature},
