+++ tests/src/common/wallet/transaction/prepare.rs
@@ -3,9 +3,9 @@
 use std::collections::HashMap;
 
 use lb_core::mantle::{
-    MantleTx, NoteId, Op, TxHash, Utxo,
+    NoteId, Op, Utxo,
     traits::Hashable as _,
-    transactions::{MantleTxBuilder, MantleTxContext},
+    transactions::{MantleTxBuilder, MantleTxContext, hash::TxHash, mantle_tx::MantleTx},
 };
 use lb_key_management_system_service::keys::ZkPublicKey;
