+++ zone-sdk/src/sequencer/zone_sequencer.rs
@@ -11,11 +11,11 @@ use lb_common_http_client::{ProcessedBlockEvent, Slot};
 use lb_core::{
     header::HeaderId,
     mantle::{
-        MantleTx, Op, SignedMantleTx,
+        Op, SignedMantleTx,
         channel::{ChannelState, SlotTimeframe, SlotTimeout},
         ops::channel::{ChannelId, MsgId, config::Keys, inscribe::Inscription},
         traits::Hashable as _,
-        transactions::{Ops, hash::TxHash, states::Unverified},
+        transactions::{Ops, hash::TxHash, mantle_tx::MantleTx, states::Unverified},
     },
 };
 use lb_key_management_system_service::keys::{Ed25519Key, Ed25519Signature};
